Flyttade relevant kod från authS till dataS
