+++ src/app/auth.service.ts
@@ -2,7 +2,6 @@ import { Injectable } from '@angular/core';
 import { HttpClient, HttpResponse, HttpErrorResponse, HttpHeaders } from '@angular/common/http';
 
 import { Router } from '@angular/router';
-import { Observable, BehaviorSubject } from 'rxjs';
 
 @Injectable({
   providedIn: 'root'
@@ -11,15 +10,13 @@ import { Observable, BehaviorSubject } from 'rxjs';
 
 
 export class AuthService {
-  mockedItems = [];
   token;
-  stream = null;
+
   constructor(private http: HttpClient, private router: Router) {
     if (localStorage.getItem('userLoggedIn')) {
       this.token = localStorage.getItem('userLoggedIn');
       console.log(this.token);
     }
-    this.stream = new BehaviorSubject(this.mockedItems);
   }
 
   login(userInput) {
@@ -36,20 +33,5 @@ export class AuthService {
   logout() {
     localStorage.removeItem('userLoggedIn');
     this.router.navigate(['/login']);
-  }
-
-  getMockedItems(resource): Observable<any> {
-    const options = {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${this.token}`
-      })
-    };
-    const itemsArray = this.http.get(resource, options);
-    itemsArray.subscribe((item: any) => {
-      this.mockedItems = item;
-      this.stream.next(this.mockedItems);
-      console.log('AuthService MockedItems', this.mockedItems);
-    });
-    return this.stream;
   }
 }
