+++ src/app/item-component/item-component.component.html
@@ -1,3 +1,3 @@
-<p *ngFor="let item of mockedItems.entries">
+<p *ngFor="let item of items.entries">
   {{ item.name  }}
 </p>
